+++ internal/datanode/io_pool.go
@@ -3,10 +3,10 @@ package datanode
 import (
 	"sync"
 
-	"github.com/milvus-io/milvus/internal/util/concurrency"
+	"github.com/milvus-io/milvus/internal/util/conc"
 )
 
-var ioPool *concurrency.Pool
+var ioPool *conc.Pool
 var ioPoolInitOnce sync.Once
 
 func initIOPool() {
@@ -15,10 +15,10 @@ func initIOPool() {
 		capacity = 32
 	}
 	// error only happens with negative expiry duration or with negative pre-alloc size.
-	ioPool = concurrency.NewPool(capacity)
+	ioPool = conc.NewPool(capacity)
 }
 
-func getOrCreateIOPool() *concurrency.Pool {
+func getOrCreateIOPool() *conc.Pool {
 	ioPoolInitOnce.Do(initIOPool)
 	return ioPool
 }
