+++ internal/querynode/impl_test.go
@@ -30,7 +30,7 @@ import (
 	"github.com/milvus-io/milvus/internal/proto/internalpb"
 	"github.com/milvus-io/milvus/internal/proto/querypb"
 	queryPb "github.com/milvus-io/milvus/internal/proto/querypb"
-	"github.com/milvus-io/milvus/internal/util/concurrency"
+	"github.com/milvus-io/milvus/internal/util/conc"
 	"github.com/milvus-io/milvus/internal/util/etcd"
 	"github.com/milvus-io/milvus/internal/util/metricsinfo"
 	"github.com/milvus-io/milvus/internal/util/sessionutil"
@@ -116,7 +116,7 @@ func TestImpl_WatchDmChannels(t *testing.T) {
 		defer func() {
 			node.taskPool = originPool
 		}()
-		node.taskPool = concurrency.NewDefaultPool()
+		node.taskPool = conc.NewDefaultPool()
 		node.taskPool.Release()
 		status, err = node.WatchDmChannels(ctx, req)
 		assert.NoError(t, err)
