+++ internal/util/conc/future.go
@@ -14,7 +14,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package concurrency
+package conc
 
 type future interface {
 	wait()
